stern: Import stern-1.26.0 as net/stern
Stern allows you to tail multiple pods on Kubernetes and multiple
containers within the pod. Each result is color coded for quicker
debugging.

The query is a regular expression or a Kubernetes resource in the form
<resource>/<name> so the pod name can easily be filtered and you don't
need to specify the exact id (for instance omitting the deployment id).
If a pod is deleted it gets removed from tail and if a new pod is added
it automatically gets tailed.

When a pod contains multiple containers Stern can tail all of them too
without having to do this manually for each one. Simply specify the
container flag to limit what containers to show. By default all
containers are listened to.

Packaged in pkgsrc-wip by myself.