bulk-small: update deps
Update the lists of stuff that gets built implicitly, not because it's
necessarily that valuable to do so but because it's worth being able to
audit just which things we're building and how expensive they are.

The following packages were implicitly included have been promoted to
explicit: pkgconf glib2. Rationale: both are extremely common, and
while glib2's expensive, if we can afford to build it in this
package's budget we probably should.

The following packages were explicitly demoted the other way: autoconf
automake. These are rarely actually needed to build things, and also
aren't expensive to build yourself. It turns out that at least one
thing we have here (screen) needs both anyway, but if it didn't
building these probably wouldn't be worthwhile.

(Neither of these sets of changes makes any functional difference.)