shells/nushell: update to 0.85.0
Themes of this release/New features
Nushell + Uutils = ❤️
he Nushell team is thrilled to announce 🥳 that we've begun working with the
uutils/coreutils team on integrating some of their foundational core utilies
into nushell. With this release, we've started with the coreutils cp command,
which we've temporarily named ucp. We're starting with ucp to allow broad
testing while the current nushell cp command remains the default.
We've already found one bug in the coreutils cp command and you might find
others. Once it stabelizes, probably with the 0.86.0 release, we'll remove the
nushell cp command and rename ucp to cp. In keeping with the nushell style,
we've only added a handful of parameters. We can, and probably will, add more
if the community determines we need them. We're so very excited and would like
to thank terts, from the coreutils team, for his excellent help and special
appreciation to dmatos2012 for the, very iterative, first uutils/coreutils
integration PR #10097 for the cp command. To read more about how this came to
be, checkout our blog post.

Read about the full list of changes at:
https://www.nushell.sh/blog/2023-09-19-nushell_0_85_0.html