sqlite3: updated to 3.43.2
Changes in this specific patch release, version 3.43.2 (2023-10-10):

Fix a couple of obscure UAF errors and an obscure memory leak.
Omit the use of the sprintf() function from the standard library in the CLI, as this now generates warnings on some platforms.
Avoid conversion of a double into unsigned long long integer, as some platforms do not do such conversions correctly.