putty: Update to 0.79
Changelog:
2023-08-26 PuTTY 0.79 released

PuTTY 0.79, released today, is mostly a bug fix release, with only
minor new features in SSH and terminal mouse handling.

The most important bug fix is that we've restored the Windows
'install scope' to the way it was in 0.77 and earlier, reverting
the security workaround we had to put into 0.78. This means the
0.79 Windows installer will not uninstall 0.78 automatically, so
we recommend uninstalling 0.78 by hand first, if you have it
installed. As before, if you've ended up with both versions installed,
uninstalling them both and then running the new installer will put
everything right.