py-hypothesis: updated to 6.88.1
6.88.1

This patch improves register_type_strategy() when used with tuple subclasses, by preventing them from being interpreted as generic and provided to strategies like st.from_type(Sequence[int]).

6.88.0

This release allows strategy-generating functions registered with register_type_strategy() to conditionally not return a strategy, by returning NotImplemented.

6.87.4

When randoms() was called with use_true_randoms=False, calling sample on it with an empty sequence and 0 elements would result in an error, when it should have returned an empty sequence to agree with the normal behaviour of random.Random. This fixes that discrepancy.

6.87.3

This patch ensures that the hypothesis codemod CLI will print a warning instead of stopping with an internal error if one of your files contains invalid syntax.

6.87.2

This patch makes some small changes to our NumPy integration to ensure forward compatibility.